add: features to the PageWrapper
