+++ src/pages/admin/payments-list/index.tsx
@@ -1,4 +1,4 @@
-import { PageWrapper } from '@/shared'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 const PaymentsListPage = () => {
   return <PageWrapper>PaymentsListPage</PageWrapper>
