+++ src/pages/admin/posts-list/index.tsx
@@ -1,4 +1,4 @@
-import { PageWrapper } from '@/shared'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 export const PostsListPage = () => {
   return <PageWrapper>PostsListPage</PageWrapper>
