+++ src/pages/admin/sign-in/index.tsx
@@ -1,4 +1,4 @@
-import { PageWrapper } from '@/shared'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 export const SignInPage = () => {
   return <PageWrapper>SignInPage</PageWrapper>
