+++ src/pages/admin/statistics/index.tsx
@@ -1,4 +1,4 @@
-import { PageWrapper } from '@/shared'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 export const StatisticsPage = () => {
   return <PageWrapper>StatisticsPage</PageWrapper>
