+++ src/pages/admin/user/[userID]/index.tsx
@@ -1,4 +1,4 @@
-import { PageWrapper } from '@/shared'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 export const UsersInfoPage = () => {
   return <PageWrapper>UsersInfoPage</PageWrapper>
