+++ src/pages/auth/create-new-password/index.tsx
@@ -1,7 +1,8 @@
 import React, { useRef } from 'react'
 
 import { CreateNewPasswordFormValues, CreatePasswordForm } from '@/feature'
-import { PageWrapper, ReturnComponent, UseFormRef, getAuthLayout, useTranslation } from '@/shared'
+import { getAuthLayout, ReturnComponent, UseFormRef, useTranslation } from '@/shared'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 const CreateNewPasswordPage = (): ReturnComponent => {
   const { t } = useTranslation()
