+++ src/pages/auth/forgot-password/index.tsx
@@ -3,14 +3,9 @@ import ReCAPTCHA from 'react-google-recaptcha'
 import { UseFormClearErrors, UseFormSetValue } from 'react-hook-form'
 
 import { ForgotPasswordForm, ForgotPasswordFormValues, SentEmailModal } from '@/feature'
-import {
-  PageWrapper,
-  RECAPTCHA_PUBLIK_KEY,
-  UseFormRef,
-  getAuthLayout,
-  useTranslation,
-} from '@/shared'
+import { getAuthLayout, RECAPTCHA_PUBLIK_KEY, UseFormRef, useTranslation } from '@/shared'
 import Script from 'next/script'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 type ForgotPasswordFormRef = UseFormRef<
   ForgotPasswordFormValues,
