+++ src/pages/auth/sign-in/index.tsx
@@ -1,15 +1,9 @@
 import { useRef } from 'react'
 
 import { SignInFormValues } from '@/feature'
-import {
-  EMPTY_STRING,
-  getAuthLayout,
-  PageWrapper,
-  ReturnComponent,
-  UseFormRef,
-  useTranslation,
-} from '@/shared'
+import { EMPTY_STRING, getAuthLayout, ReturnComponent, UseFormRef, useTranslation } from '@/shared'
 import dynamic from 'next/dynamic'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 const DynamicSignInForm = dynamic(
   import('../../../feature/auth/ui/sign-in-form').then(module => module.SignInForm)
