+++ src/pages/settings/index.tsx
@@ -1,5 +1,5 @@
 import { getBaseAppLayout } from '@/shared/layouts/base-app-layout'
-import { PageWrapper } from '@/shared/layouts/page-wrapper'
+import { PageWrapper } from '@/widgets/page-wrapper'
 
 const SettingsPage = () => {
   return (
