+++ src/shared/layouts/base-app-layout.tsx
@@ -6,7 +6,7 @@ import { Main } from '@/widgets/main/main'
 import { Sidebar } from '@/widgets/sidebar'
 import { NextPage } from 'next'
 
-import { LayoutContextProvider } from '@/shared/lib/context/layout-context/layout-context'
+import { LayoutContextProvider } from '../lib/context/layout-context'
 
 /** Общий Лайаут для всех страниц */
 export const BaseAppLayout: NextPage<PropsWithChildren> = ({ children }): ReturnComponent => {
